driver: Move driver store functions to their own file
